Move the Checkpoints library to `utils/structs`
