+++ contracts/governance/utils/Votes.sol
@@ -5,8 +5,8 @@ pragma solidity ^0.8.0;
 import "../../interfaces/IERC5805.sol";
 import "../../utils/Context.sol";
 import "../../utils/Nonces.sol";
-import "../../utils/Checkpoints.sol";
 import "../../utils/cryptography/EIP712.sol";
+import "../../utils/structs/Checkpoints.sol";
 
 /**
  * @dev This is a base abstract contract that tracks voting units, which are a measure of voting power that can be
