+++ contracts/token/ERC721/extensions/ERC721Consecutive.sol
@@ -5,8 +5,8 @@ pragma solidity ^0.8.1;
 
 import "../ERC721.sol";
 import "../../../interfaces/IERC2309.sol";
-import "../../../utils/Checkpoints.sol";
 import "../../../utils/structs/BitMaps.sol";
+import "../../../utils/structs/Checkpoints.sol";
 
 /**
  * @dev Implementation of the ERC2309 "Consecutive Transfer Extension" as defined in
