+++ contracts/utils/structs/Checkpoints.sol
@@ -1,11 +1,11 @@
 // SPDX-License-Identifier: MIT
-// OpenZeppelin Contracts (last updated v4.8.1) (utils/Checkpoints.sol)
+// OpenZeppelin Contracts (last updated v4.8.1) (utils/structs/Checkpoints.sol)
 // This file was procedurally generated from scripts/generate/templates/Checkpoints.js.
 
 pragma solidity ^0.8.0;
 
-import "./math/Math.sol";
-import "./math/SafeCast.sol";
+import "../math/Math.sol";
+import "../math/SafeCast.sol";
 
 /**
  * @dev This library defines the `History` struct, for checkpointing values as they change at different points in
