+++ scripts/generate/templates/Checkpoints.t.js
@@ -7,8 +7,8 @@ const header = `\
 pragma solidity ^0.8.0;
 
 import "forge-std/Test.sol";
-import "../../contracts/utils/Checkpoints.sol";
 import "../../contracts/utils/math/SafeCast.sol";
+import "../../contracts/utils/structs/Checkpoints.sol";
 `;
 
 /* eslint-disable max-len */
